+++ Src/AppLib/VpnHood.AppLib.App/AppOptions.cs
@@ -36,10 +36,8 @@ public class AppOptions(string appId, string storageFolderName, bool isDebugMode
     public ITrackerFactory? TrackerFactory { get; set; }
     public TimeSpan ReconnectTimeout { get; set; } = ClientOptions.Default.ReconnectTimeout;
     public TimeSpan AutoWaitTimeout { get; set; } = ClientOptions.Default.AutoWaitTimeout;
-
     public bool? LogAnonymous { get; set; } =
         isDebugMode ? false : null; // it follows user's settings if it set to null
-
     public TimeSpan ServerQueryTimeout { get; set; } = ClientOptions.Default.ServerQueryTimeout;
     public TimeSpan ConnectTimeout { get; set; } = Debugger.IsAttached ? Timeout.InfiniteTimeSpan : TimeSpan.FromSeconds(60);
     public bool AutoDiagnose { get; set; } = true;
@@ -51,4 +49,5 @@ public class AppOptions(string appId, string storageFolderName, bool isDebugMode
     public TimeSpan? EventWatcherInterval { get; set; } // set if you don't call State periodically
     public bool DisconnectOnDispose { get; set; }
     public LogOptions LogOptions { get; set; } = new();
+    public bool AdjustForSystemBars { get; set; } = true;
 }