+++ Src/Core/VpnHood.Core.Client.Device/ActiveUiContext.cs
@@ -2,6 +2,7 @@
 
 namespace VpnHood.Core.Client.Device;
 
+// todo: rename to CurrentUiContext
 public static class ActiveUiContext
 {
     private static IUiContext? _context;