+++ Src/Core/VpnHood.Core.Client.Device/IDevice.cs
@@ -6,9 +6,11 @@ public interface IDevice : IAsyncDisposable
     bool IsExcludeAppsSupported { get; }
     bool IsIncludeAppsSupported { get; }
     bool IsAlwaysOnSupported { get; }
+    bool IsBindProcessToVpnSupported { get; }
     string OsInfo { get; }
     DeviceMemInfo? MemInfo { get; }
     DeviceAppInfo[] InstalledApps { get; }
     Task RequestVpnService(IUiContext? uiContext, TimeSpan timeout, CancellationToken cancellationToken);
     Task StartVpnService(CancellationToken cancellationToken);
+    void BindProcessToVpn(bool value);
 }